+++ Source/Gorgon/Types.h
@@ -15,7 +15,7 @@ namespace Gorgon {
 	typedef float Float;
 #endif
     
-    static const Float PI = 3.141592653589793;
+    static const Float PI = (Float)3.141592653589793;
 	
 	/// A class that has no members and can be used as placeholder
 	class Empty {};
